+++ includes/platformer-engine/Builder/ObjectBuilder.hpp
@@ -1,6 +1,8 @@
 #ifndef PLATFORMER_ENGINE_BUILDER_HPP
 #define PLATFORMER_ENGINE_BUILDER_HPP
 
+#include "Texture/AnimatedSprite.hpp"
+
 /**
  * @brief Base class for GameObjectBuilders.
 */
@@ -10,37 +12,46 @@ public:
      * @brief Adds an AudioSource component to the current GameObject being build.
      * @return reference to ObjectBuilder, allows method chaining.
      */
-    virtual ObjectBuilder& AddAudioSource() = 0;
+    virtual auto AddAudioSource() -> ObjectBuilder & = 0;
 
     /**
      * @brief Adds an Animator component to the current GameObject being build.
+     * @param animatedSprite the AnimatedSprite to be used by the Animator.
      * @return reference to ObjectBuilder, allows method chaining.
      */
-    virtual ObjectBuilder& AddAnimator() = 0;
+    virtual auto AddAnimator(std::shared_ptr<platformer_engine::AnimatedSprite> animatedSprite) -> ObjectBuilder & = 0;
+
+    /**
+     * @brief Adds an Animator component to the current GameObject being build.
+     * @param animatedSprite List of AnimatedSprite to be used by the Animator.
+     * @return reference to ObjectBuilder, allows method chaining.
+     */
+    virtual auto
+    AddAnimator(std::vector<std::shared_ptr<platformer_engine::AnimatedSprite>> animatedSprite) -> ObjectBuilder & = 0;
 
     /**
      * @brief Adds a BehaviourScript component to the current GameObject being build.
      * @return reference to ObjectBuilder, allows method chaining.
      */
-    virtual ObjectBuilder& AddBehaviourScript() = 0;
+    virtual auto AddBehaviourScript() -> ObjectBuilder & = 0;
 
     /**
      * @brief Adds a Collider component to the current GameObject being build.
      * @return reference to ObjectBuilder, allows method chaining.
      */
-    virtual ObjectBuilder& AddCollider() = 0;
+    virtual auto AddCollider() -> ObjectBuilder & = 0;
 
     /**
      * @brief Adds a RigidBody component to the current GameObject being build.
      * @return reference to ObjectBuilder, allows method chaining.
      */
-    virtual ObjectBuilder& AddRigidBody() = 0;
+    virtual auto AddRigidBody() -> ObjectBuilder & = 0;
 
     /**
      * @brief Adds a Sprite component to the current GameObject being build.
      * @return reference to ObjectBuilder, allows method chaining.
      */
-    virtual ObjectBuilder& AddSprite() = 0;
+    virtual auto AddSprite(const std::shared_ptr<spic::Sprite> &sprite) -> ObjectBuilder & = 0;
 };
 
 #endif //PLATFORMER_ENGINE_BUILDER_HPP