+++ src/Texture/TextureManager.cpp
@@ -8,19 +8,19 @@ auto platformer_engine::TextureManager::LoadTexture(const std::string &id, const
 
 void
 platformer_engine::TextureManager::DrawTexture(const std::string &id, int x, int y, int width, int height,
-                                               const SPIC_RendererFlip &flip) {
-    _graphicsFacade->DrawTexture(id, x, y, width, height, flip);
+                                               const SPIC_RendererFlip &flip, double scale) {
+    _graphicsFacade->DrawTexture(id, x, y, width, height, flip, scale);
 }
 
 void platformer_engine::TextureManager::DrawTile(const std::string &tileSetID, int tileSize, int x, int y, int row, int frame,
-                                                 const platformer_engine::SPIC_RendererFlip &flip) {
-    _graphicsFacade->DrawTile(tileSetID, tileSize, x, y, row, frame, flip);
+                                                 const platformer_engine::SPIC_RendererFlip &flip, double scale) {
+    _graphicsFacade->DrawTile(tileSetID, tileSize, x, y, row, frame, flip, scale);
 }
 
 void
 platformer_engine::TextureManager::DrawFrame(const std::string &id, int x, int y, int width, int height, int row, int frame,
-                                             const platformer_engine::SPIC_RendererFlip &flip) {
-    _graphicsFacade->DrawFrame(id, x, y, width, height, row, frame, flip);
+                                             const platformer_engine::SPIC_RendererFlip &flip, double scale) {
+    _graphicsFacade->DrawFrame(id, x, y, width, height, row, frame, flip, scale);
 }